Changed keys/labels for resticted term level/score. Moved the definition from a component file to the model file for better organization.
